Create initial ViewEventsView, and write tests for its GET method
ViewEventsView for rendering the search and view events section template
of a user's homepage was created. The GET method retrieves events that
a user has registered their interest in or that they have upcoming
as a confirmed attendee. These events form part of the context of the
section template. Tests written to test the GET method logic and
template rendering.