Add listeners to withdraw buttons. Create withdrawFromEventFetchHandler
Click event listeners added to the event withdrawal buttons. Events
trigger fetch requests to the post method of the ViewEventsView via
the created withdrawFromEventFetchHandler. When successful a user
is withdrawn from an event they are attending or are interested in,
and the event is no longer visible to the user in their search_events
section of their homepage. Tests written for the fetch
requests.