Add listeners to register buttons. Create registerInterestFetchHandler
Click event listeners added to the register interest buttons.
registerInterestFetchHandler created that carries out POST fetch request
to the SearchAdvertsView, in order to attempt to register interest
in an event for a user. Sweetalert2 alert feedback given to user
to indicate success of request. Event advert hidden if the request
is successful.